refactor: remove tinyobjloader and move resource systems to resources dir
