+++ src/Core/Renderer/SceneSystem/Node.cc
