+++ src/Core/Resources/Loaders/TextLoader.hh
@@ -3,7 +3,7 @@
 
 #include "esppch.hh"
 
-#include "Core/Systems/ResourceSystem.hh"
+#include "Core/Resources/Systems/ResourceSystem.hh"
 
 namespace esp
 {
