+++ src/Core/Resources/Systems/MaterialSystem.cc
