+++ src/Core/Resources/Systems/MaterialSystem.hh
