+++ src/Core/Resources/Systems/ResourceSystem.hh
