+++ src/Core/Resources/Systems/TextureSystem.hh
