+++ src/Platform/Vulkan/Resources/VulkanTexture.hh
@@ -8,7 +8,7 @@
 #include "VulkanSampler.hh"
 
 // Core
-#include "Core/Systems/TextureSystem.hh"
+#include "Core/Resources/Systems/TextureSystem.hh"
 
 namespace esp
 {
