+++ src/Platform/Vulkan/VulkanResourceManager.hh
@@ -3,7 +3,7 @@
 
 #include "esppch.hh"
 
-#include "Core/Systems/TextureSystem.hh"
+#include "Core/Resources/Systems/TextureSystem.hh"
 
 namespace esp
 {
